fixing loading bar

+++ interface/src/octree/SafeLanding.h
@@ -50,7 +50,7 @@ private:
     static constexpr int INVALID_SEQUENCE = -1;
     int _initialStart { INVALID_SEQUENCE };
     int _initialEnd { INVALID_SEQUENCE };
-    int _maxTrackedEntityCount { 0 };
+    float _maxTrackedEntityCount { 0.0f };
 
     struct SequenceLessThan {
         bool operator()(const int& a, const int& b) const;
